+++ lib/screens/image_picker/selector/local_selector.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:image_picker/image_picker.dart';
 
 import '../../../lang/lang.dart';
@@ -65,13 +66,14 @@ class _LocalImageSelectorState extends State<LocalImageSelector>
           ),
         ),
         ColorInfo(
+          shrinkable: false,
           color: color,
           onExpand: () {
             pickerKey.currentState?.loadSnapshotBytes();
             setState(() {});
           },
           leading: buildCompactIconButton(
-            icon: Icon(Icons.photo_album_rounded),
+            icon: FaIcon(FontAwesomeIcons.image),
             tooltip: lang.selectPhoto,
             onPressed: getImage,
           ),